Implemented GLFW event handlers and connected it to the event system.
Windows can close now! ;)

+++ include/HazelPVR/Events/MouseEvent.h
@@ -6,7 +6,7 @@ namespace HazelPVR {
 
 	class MouseMovedEvent : public Event {
 	public:
-		MouseMovedEvent(float x, float y)
+		MouseMovedEvent(double x, double y)
 			: m_MouseX(x), m_MouseY(y) {}
 
 		inline float GetX() const { return m_MouseX; }
@@ -26,7 +26,7 @@ namespace HazelPVR {
 
 	class MouseScrolledEvent : public Event	{
 	public:
-		MouseScrolledEvent(float xOffset, float yOffset)
+		MouseScrolledEvent(double xOffset, double yOffset)
 			: m_XOffset(xOffset), m_YOffset(yOffset) {}
 
 		inline float GetXOffset() const { return m_XOffset; }
